refactor(constants): move custom view events names to constants package
