--- packages/application-components/src/components/custom-views/custom-view-loader/constants.ts
@@ -1,9 +1,0 @@
-export const CUSTOM_VIEWS_EVENTS_NAMES = {
-  CUSTOM_VIEW_BOOTSTRAP: 'custom-view-bootstrap',
-  CUSTOM_VIEW_INITIALIZATION: 'custom-view-initialization',
-};
-
-export const CUSTOM_VIEWS_EVENTS_META = {
-  SOURCE: 'mc-host-application',
-  DESTINATION_PREFIX: 'custom-view-',
-};
