+++ packages/application-components/src/components/custom-views/custom-view-loader/custom-view-loader.tsx
@@ -4,16 +4,14 @@ import { useIntl } from 'react-intl';
 import { useShowNotification } from '@commercetools-frontend/actions-global';
 import { useApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 import {
+  CUSTOM_VIEWS_EVENTS_NAMES,
+  CUSTOM_VIEWS_EVENTS_META,
   DOMAINS,
   NOTIFICATION_KINDS_PAGE,
 } from '@commercetools-frontend/constants';
 import { reportErrorToSentry } from '@commercetools-frontend/sentry';
 import { TCustomView } from '../../../types/generated/settings';
 import CustomPanel from '../custom-panel/custom-panel';
-import {
-  CUSTOM_VIEWS_EVENTS_NAMES,
-  CUSTOM_VIEWS_EVENTS_META,
-} from './constants';
 import messages from './messages';
 
 type TCustomViewIframeMessage = {
