--- packages/application-shell/src/components/custom-view-shell/constants.ts
@@ -1,10 +1,0 @@
-// TODO: This file is duplicated (also exists in application-components custom views components)
-export const CUSTOM_VIEWS_EVENTS_NAMES = {
-  CUSTOM_VIEW_BOOTSTRAP: 'custom-view-bootstrap',
-  CUSTOM_VIEW_INITIALIZATION: 'custom-view-initialization',
-};
-
-export const CUSTOM_VIEWS_EVENTS_META = {
-  SOURCE: 'mc-host-application',
-  DESTINATION_PREFIX: 'custom-view-',
-};
