+++ packages/constants/src/constants.ts
@@ -209,3 +209,13 @@ export const HTTP_SECURITY_HEADERS = {
   'X-Frame-Options': 'DENY',
   'Referrer-Policy': 'same-origin',
 } as const;
+
+// Custom Views events (messages sent between the host application and the custom view)
+export const CUSTOM_VIEWS_EVENTS_NAMES = {
+  CUSTOM_VIEW_BOOTSTRAP: 'custom-view-bootstrap',
+  CUSTOM_VIEW_INITIALIZATION: 'custom-view-initialization',
+};
+export const CUSTOM_VIEWS_EVENTS_META = {
+  SOURCE: 'mc-host-application',
+  DESTINATION_PREFIX: 'custom-view-',
+};
